downstream_tasks/requirements.txt
bayesian-optimization==1.4.1
botorch==0.12.0
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.10.0
gpytorch==1.13
importlib-metadata==8.5.0
jaxtyping==0.2.19
jinja2==3.1.4
joblib==1.4.2
linear_operator==0.5.3
loguru==0.7.2
markupsafe==3.0.2
mpmath==1.3.0
multipledispatch==0.6.0
networkx==3.4.2
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-tzdata==2024.2
pytorch==2.5.1
scikit-learn==1.3.0
scipy==1.14.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typeguard==4.4.1
typing-extensions==4.12.2
xgboost==2.1.3
zipp==3.21.0